Fix initial state of form components
Before, we were implicitly initializing the `TextField` component with
`undefined`, which results in an uncontrolled form component being
rendered at first. Instead, we specify an empty string in the initial
component state, so it is controlled the entire time. We can then reset
it by setting the same state again.